requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
DAWG-Python==0.7.2
docopt==0.6.2
frozenlist==1.4.1
idna==3.7
langdetect==1.0.9
magic-filter==1.0.12
multidict==6.0.5
pyaes==1.6.1
pyasn1==0.6.0
pydantic==2.7.4
pydantic_core==2.18.4
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyspellchecker==0.8.1
python-dotenv==1.0.1
rsa==4.9
six==1.16.0
Telethon==1.36.0
typing_extensions==4.12.2
yarl==1.9.4
